requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bitarray==3.4.1
cdp-sdk==1.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
coinbase-advanced-py==1.8.2
construct==2.10.68
construct-typing==0.5.6
cryptography==45.0.2
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.6.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonalias==0.1.1
multidict==6.4.4
numpy==2.2.6
pandas==2.2.3
parsimonious==0.10.0
propcache==0.3.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
six==1.17.0
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
toolz==1.0.0
types-requests==2.32.0.20250515
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
web3==7.10.0
websockets==13.1
yarl==1.20.0
